simplified PVA and re-ran survival analysis and projection
IPM did not converge and survival estimates are now dependent on longitude, which results in awkward need to specify from which longitude the juveniles emerge. Abandoned the IPM and created simple PVA but got stuck at what longitude value to specify as 25.5 leads to unrealistic lambda (1.08) whereas 24.3. leads to realistic lambda (1.006) but unrealistic juvenile survival (5-20%).